media-source/SourceBuffer-abort-readyState.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>SourceBuffer#abort() if SourceBuffer#updating attribute equals true</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>
    <video id='v' controls autoplay width='320' height='240'></video>
</div>
<div id="log"></div>

<script>
var FILE = '/media/white.webm';
var video = document.querySelector('video');
var mime = 'video/webm; codecs="vorbis,vp8"';

//check the browser supports the MIME used in this test
function canPlayType(mime) {
    if(video.canPlayType(mime) != 'probably') {
        this.step(function() {
            assert_unreached("Browser doesn't support the MIME used in this test.");
        });
        this.done();
        return false;
    }
    return true;
}
function GET(url, processBody) {
    var xhr = new XMLHttpRequest();
    xhr.open('GET', url, true);
    xhr.responseType = 'arraybuffer';
    xhr.send();
    xhr.onload = function(e) {
        if (xhr.status != 200) {
            alert("Unexpected status code " + xhr.status + " for " + url);
            return false;
        }
        processBody(new Uint8Array(xhr.response));
    };
}
async_test(function(t) {
    if(!canPlayType.bind(t)(mime)) {
        return;
    }
    GET(FILE, function(data) {
        var mediaSource = new MediaSource();
        var events = [];
        mediaSource.addEventListener('sourceopen', function(e){
            var sourceBuffer = mediaSource.addSourceBuffer(mime);
            sourceBuffer.addEventListener('updatestart', function(e){
                sourceBuffer.abort();
                t.step(function() {
                    assert_equals(sourceBuffer.updating, false);
                });
            });
            sourceBuffer.addEventListener('abort', function(e){
                events.push('abort');
            });
            sourceBuffer.addEventListener('updateend', function(e){
                events.push('updateend');
                t.step(function() {
                    assert_array_equals(events, ['abort', 'updateend']);
                });
                t.done();
            });
            sourceBuffer.appendBuffer(data);
        });
        video.src = window.URL.createObjectURL(mediaSource);
    });
}, 'SourceBuffer#abort() : in case the updating attribute equals true');

</script>
</body>
</html>
